[fix] address issues mentioned for PRs components
